Change to bit-mag fixed-point representation

+++ core/include/core/Simulator.h
@@ -27,12 +27,10 @@ namespace core {
         cnpy::Array<T> adjustPadding(const cnpy::Array<T> &array, int padding);
 
         /* Return the optimal encoding for the given value
-         * @param value     Value we want to encode
-         * @param mag       Magnitude of the value
-         * @param prec      Precision of the value
+         * @param value     Value we want to encode WITHOUT the sign
          * @return          Value with the optimal encoding
          */
-        uint32_t booth_encoding(uint16_t value);
+        uint16_t booth_encoding(uint16_t value);
 
     };
